+++ src/main/java/com/a5lab/tabr/domain/radars/RadarService.java
@@ -23,6 +23,4 @@ public interface RadarService {
   Radar save(Radar radar);
 
   void deleteById(Long id);
-
-  Radar createRadarEnv(Radar radar);
 }
