+++ src/main/java/com/a5lab/tabr/domain/wizard/WizardService.java
@@ -1,0 +1,8 @@
+package com.a5lab.tabr.domain.wizard;
+
+import com.a5lab.tabr.domain.radars.Radar;
+
+public interface WizardService {
+
+  Radar createRadar(Radar radar);
+}
